+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/capacitations/DisabilitiesList.kt
@@ -45,6 +45,8 @@ import com.app.arcabyolimpo.presentation.ui.components.atoms.buttons.AddButton
 import com.app.arcabyolimpo.presentation.navigation.Screen
 import com.app.arcabyolimpo.presentation.ui.components.atoms.inputs.SearchInput
 import com.app.arcabyolimpo.presentation.ui.components.molecules.BeneficiaryCard
+import com.app.arcabyolimpo.presentation.ui.components.atoms.alerts.Snackbarcustom
+import com.app.arcabyolimpo.presentation.ui.components.molecules.DisabilityCard
 import com.app.arcabyolimpo.ui.theme.ArcaByOlimpoTheme
 import com.app.arcabyolimpo.ui.theme.Background
 import kotlinx.coroutines.launch
@@ -101,6 +103,7 @@ fun DisabilitiesListScreen(
         state = state,
         snackbarHostState = snackbarHostState,
         onSearchTextChange = viewModel::onSearchTextChange,
+        beneficiaryImage = state.beneficiary?.image,
         onDisabilityClick = onDisabilityClick,
         onAddClick = onAddClick,
         onBackClick = onBackClick,
@@ -116,6 +119,7 @@ fun DisabilitiesList(
     onSearchTextChange: (String) -> Unit,
     onDisabilityClick: (String) -> Unit,
     onAddClick: () -> Unit,
+    beneficiaryImage: String?,
     onBackClick: () -> Unit,
     navController: NavHostController
 ) {
@@ -218,7 +222,10 @@ fun DisabilitiesList(
                         else -> {
                             LazyVerticalGrid(
                                 columns = GridCells.Fixed(2),
-                                contentPadding = PaddingValues(vertical = 16.dp),
+                                contentPadding = PaddingValues(
+                                    horizontal = 16.dp,
+                                    vertical = 16.dp
+                                ),
                                 horizontalArrangement = Arrangement.spacedBy(8.dp),
                                 verticalArrangement = Arrangement.spacedBy(8.dp),
                                 modifier =
@@ -227,13 +234,13 @@ fun DisabilitiesList(
                                         .weight(1f),
                             ) {
                                 items(state.disabilities, key = { it.id }) { disability ->
-                                    BeneficiaryCard(
+                                    DisabilityCard(
                                         name = disability.name,
                                         onClick = {
-                                            navController.navigate(Screen.DisabilityDetail.createRoute(disability.id))},
+                                            navController.navigate(Screen.DisabilityDetail.createRoute(disability.id))
+                                        },
                                         cardModifier = Modifier
                                             .fillMaxWidth()
-                                            .padding(horizontal = 8.dp)
                                             .height(160.dp),
                                         contentPadding = PaddingValues(
                                             vertical = 20.dp,